# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Simversity API
  version: 0.1.0
  description: AI-powered teaching simulator API for practicing responses to student misconceptions.
  contact:
    name: Simversity
    url: https://simversity.org
  license:
    name: MIT
servers:
  - url: https://simversity.org
    description: Production
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/claim-role:
    post:
      tags: [Auth]
      summary: Claim role via invite code
      description: Redeem an invite code to upgrade user role to teacher or admin.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [inviteCode]
              properties:
                inviteCode:
                  type: string
                  minLength: 1
                  maxLength: 200
      responses:
        "200":
          description: Role claimed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  role:
                    type: string
                    enum: [teacher, admin]
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: Invalid or expired code
        "409":
          description: Code already used
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/courses:
    get:
      tags: [Courses]
      summary: List published courses
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated course list
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: "#/components/schemas/CourseSummary"
                  total:
                    type: integer
...
  /api/conversations:
    get:
      tags: [Conversations]
      summary: List user conversations
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, abandoned]
        - name: search
          in: query
          description: Filter by scenario title (partial match)
          schema:
            type: string
        - name: from
          in: query
          description: Filter conversations started on or after this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Filter conversations started on or before this date (YYYY-MM-DD)
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated conversation list
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversations:
                    type: array
                    items:
                      $ref: "#/components/schemas/ConversationSummary"
                  total:
                    type: integer
    post:
      tags: [Conversations]
      summary: Start new conversation
      description: Creates a new conversation for a scenario and returns opening messages from student agents.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [scenarioId]
              properties:
                scenarioId:
                  type: string
                  format: uuid
      responses:
        "201":
          description: Conversation created with opening messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversation:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      userId:
                        type: string
                        format: uuid
                      scenarioId:
                        type: string
                        format: uuid
                      status:
                        type: string
                        enum: [active]
                      messageCount:
                        type: integer
                  messages:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        conversationId:
                          type: string
                          format: uuid
                        role:
                          type: string
                          enum: [assistant]
                        content:
                          type: string
                        agentId:
                          type: string
                          format: uuid
                          nullable: true
                        sortOrder:
                          type: integer
                        agentName:
                          type: string
                          nullable: true
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/progress:
    get:
      tags: [Progress]
      summary: Get user progress
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated progress records
          content:
            application/json:
              schema:
                type: object
                properties:
                  progress:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProgressRecord"
                  total:
                    type: integer
...
  /api/models:
    get:
      tags: [Models]
      summary: List available AI models
      security:
        - session: []
      responses:
        "200":
          description: Available models from NEAR AI Cloud
          content:
            application/json:
              schema:
                type: object
                properties:
                  models:
                    type: array
                    items:
                      $ref: "#/components/schemas/ModelInfo"
...
  /api/budget:
    get:
      tags: [User]
      summary: Get daily message budget status
      security:
        - session: []
      responses:
        "200":
          description: Budget status
          content:
            application/json:
              schema:
                type: object
                properties:
                  used:
                    type: integer
                  limit:
                    type: integer
                  remaining:
                    type: integer
                  enabled:
                    type: boolean
...
  /api/admin/courses:
    get:
      tags: [Admin - Courses]
      summary: List all courses (admin)
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
      responses:
        "200":
          description: Paginated course list
          content:
            application/json:
              schema:
                type: object
                properties:
                  courses:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminCourse"
                  total:
                    type: integer
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Admin - Courses]
      summary: Create course
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Course created
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by Better-Auth on sign-in
  parameters:
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
    CourseSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        gradeLevel:
          type: string
        subject:
          type: string
        scenarioCount:
          type: integer
        visibility:
          type: string
          enum: [private, shared, published, archived]
        createdBy:
          type: string
          format: uuid
          nullable: true
    ConversationSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        scenarioId:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, completed, abandoned]
        messageCount:
          type: integer
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
        scenarioTitle:
          type: string
        studentName:
          type: string
          nullable: true
        courseId:
          type: string
          format: uuid
...
    CourseInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        gradeLevel:
          type: string
          maxLength: 100
        subject:
          type: string
          maxLength: 100
        visibility:
          type: string
          enum: [private, shared, published, archived]
...
    AdminCourse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        gradeLevel:
          type: string
        subject:
          type: string
        scenarioCount:
          type: integer
        visibility:
          type: string
          enum: [private, shared, published, archived]
        createdBy:
          type: string
          format: uuid
          nullable: true
        updatedBy:
          type: string
          format: uuid
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProgressRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        scenarioId:
          type: string
          format: uuid
        status:
          type: string
          enum: [not_started, in_progress, completed]
        latestConversationId:
          type: string
          format: uuid
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ModelInfo:
      type: object
      properties:
        id:
          type: string
          description: Model ID (e.g. "deepseek-ai/DeepSeek-V3.1")
        label:
          type: string
          description: Display name (e.g. "DeepSeek V3.1")
        context:
          type: string
          description: Context window size (e.g. "128K")
        tier:
          type: string
          enum: ["$", "$$", "$$$"]
          description: Pricing tier
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"